backend/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
elasticsearch==7.0.2
celery==4.3.0
redis==3.3.11
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
psycopg2==2.8.4
scipy==1.3.0
statsmodels==0.10.0
stldecompose==0.0.5

